LayoutTests/imported/w3c/web-platform-tests/html/semantics/grouping-content/the-ol-element/ol.start-reflection-2-expected.txt

Three
Two
One

FAIL ol.start - reflection test assert_equals: expected 3 but got 0
